Move ViconReceiver class to separate library
