+++ include/vicon_transformer/errors.hpp
@@ -8,7 +8,7 @@
 #include <fmt/format.h>
 #include <vicon-datastream-sdk/DataStreamClient.h>
 
-#include <vicon_transformer/ostream.hpp>
+#include <vicon_transformer/fmt.hpp>
 
 namespace vicon_transformer
 {
